refactor: separate coinext module from the websocket
