--- jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'node',
-  rootDir: 'src',
-  testMatch: ["<rootDir>/**/*.test.(js|ts)"]
-};
