+++ src/config.ts
@@ -1,5 +1,6 @@
 export default {
   API_V2_URL: 'wss://api.coinext.com.br/WSGateway/',
+  API_TEST_URL: 'ws://localhost:8080',
   websocketPing: false,
   AP_KEY: '',
   AP_SECRET: '',
